# remove inspect.stack() from @rule parsing to fix import time regression in py3