--- a/tests/python/pants_test/engine/test_rules.py
+++ b/tests/python/pants_test/engine/test_rules.py
@@ -60,6 +60,9 @@ class SubA(A):
 _suba_root_rules = [RootRule(SubA)]
 
 
+_this_is_not_a_type = 3
+
+
 @console_rule('example', [Select(Console)])
 def a_console_rule_generator(console):
   a = yield Get(A, str('a str!'))
@@ -81,7 +84,7 @@ Rule entry A() had an unexpected type: <class 'pants_test.engine.test_rules.A'>.
       RuleIndex.create([A()])
 
 
-class RuleGraphTest(unittest.TestCase):
+class RuleGraphTest(TestBase):
   def test_ruleset_with_missing_product_type(self):
     @rule(A, [Select(B)])
     def a_from_b_noop(b):
@@ -694,7 +697,24 @@ class RuleGraphTest(unittest.TestCase):
                      }""").strip(),
                                     subgraph)
 
-  def test_validate_yield_statements(self):
+  def test_invalid_get_arguments(self):
+    with self.assertRaisesWithMessage(ValueError, """\
+Could not resolve type `XXX` in top level of module pants_test.engine.test_rules"""):
+      class XXX(object): pass
+      @rule(A, [])
+      def f():
+        a = yield Get(A, XXX, 3)
+        yield a
+
+    # This fails because the argument is defined in this file's module, but it is not a type.
+    with self.assertRaisesWithMessage(ValueError, """\
+Expected a `type` constructor, but got: _this_is_not_a_type"""):
+      @rule(A, [])
+      def g():
+        a = yield Get(A, _this_is_not_a_type, 3)
+        yield a
+
+  def test_validate_yield_statements_in_rule_body(self):
     with self.assertRaisesRegexp(_RuleVisitor.YieldVisitError, re.escape('yield A()')):
       @rule(A, [])
       def f():