@cosmicexplorer 2019-03-28
@Eric-Arellano I believe I have addressed both of your suggestions!